+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Hunter/StranglethornHeart.cs
@@ -7,4 +7,10 @@ public class StranglethornHeart: ResurrectionCard
 	protected override bool FilterCard(Card card) => card.IsBeast() && card.Cost >= 5;
 
 	protected override bool ResurrectsMultipleCards() => true;
+
+	public override bool ShouldShowForOpponent(Player opponent)
+	{
+		var card = Database.GetCardFromId(GetCardId());
+		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentFormat, opponent.OriginalClass) && GetRelatedCards(opponent).Count > 1;
+	}
 }