+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## **Unreleased**
+**Hearthstone**:
+- Fixed some related cards appearing when they are not relevant.
+
 ## **Release v1.38.1 - 2024-12-25**
 **Hearthstone**:
 - Fixed some Year of the Wolf sets not being considered as Wild.